estorehouse/storehouse/urls.py
from django.urls import include, path

from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView

from rest_framework.routers import DefaultRouter

from storehouse import views


# Create a router and register our viewsets with it.
router = DefaultRouter()
router.register(r"product", views.ProductViewSet, basename="product")
router.register(r"order", views.OrderViewSet, basename="order")
router.register(r"orderitems", views.OrderItemViewSet, basename="orderitem")
router.register(r"users", views.UserViewSet, basename="user")

# The API URLs are now determined automatically by the router.
urlpatterns = [
    path("", include(router.urls)),
]


urlpatterns += [
    path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
    # Optional UI:
    path("api/schema/swagger-ui/", SpectacularSwaggerView.as_view(url_name="schema"), name="swagger-ui"),
    path("api/schema/redoc/", SpectacularRedocView.as_view(url_name="schema"), name="redoc"),
]
