estorehouse/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
appnope==0.1.3
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
billiard==3.6.4.0
black==23.1.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==3.0.0
Django==4.1.7
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
drf-spectacular==0.26.1
executing==1.2.0
Faker==18.3.0
filelock==3.10.2
flake8==5.0.4
flake8-bugbear==23.2.13
flake8-builtins==2.1.0
flake8-django==1.1.5
flake8-import-order==0.18.2
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-return==1.2.0
flake8-spellcheck==0.28.0
frozenlist==1.3.3
html5lib==1.1
httpie==3.2.1
httpio==0.3.0
idna==3.4
inflection==0.5.1
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.14.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.6.0
pypng==0.20220715.0
pyproject_api==1.5.1
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.3
reportlab==3.6.12
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
tox==4.4.7
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.9
yarl==1.8.2
